gatus/pyproject.toml
[project]
name = "gatus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.45.0",
    "openai>=2.7.1",
    "python-dotenv>=1.1.1",
    "tabulate>=0.9.0",
    "tinydb>=4.8.2",
    "typer>=0.19.2",
]
